requirements.txt
Flask
google-genai